requirements/development.txt
# Зависимости для разработки
-r base.txt

# Инструменты для разработки
django-debug-toolbar>=4.2.0,<5.0.0
django-extensions>=3.2.0,<4.0.0

# Инструменты для качества кода
flake8>=6.1.0,<7.0.0
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0

# Инструменты для безопасности
safety>=3.0.0,<4.0.0
bandit>=1.7.0,<2.0.0 